Remove dependency of ImGui from JadeEngine (UNSTABLE BUILD)
I completely removed the dependency of ImGui from the engine. Since the engine
has no concept of a 'level editor' I figured it was right to move that dependency
over to the editor application, and this has helped code clarity **tremendously**.

+++ JadeEngine/include/jade/renderer/RenderBatch.h
@@ -11,7 +11,7 @@ namespace Jade
         glm::vec4 color;
         glm::vec2 texCoords;
         float texId;
-        float entityId;
+        uint32 entityId;
     };
 
     class RenderBatch
